app/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-multipart
Pillow
qrcode
python-barcode
celery
redis
